backend/requirements.txt
# YouTube AutoPilot Hub - Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# Google APIs
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Progress tracking
tqdm==4.66.1
